[ refactor ] separate banner image file
